creates pruning model
